clean refactor move funtions
